07/07 refactor(analysis): github 분석 및 훅 부분 변경, 파일구조 변경
