+++ src/github/github-webhook.controller.ts
@@ -1,62 +1,30 @@
-import { Body, Controller, Get, Param, Post, Query } from '@nestjs/common';
+import { Body, Controller, Get, Param, Post, Query, Headers } from '@nestjs/common';
 import { GithubWebhookService } from './github-webhook.service';
-
+import { AnalysisService } from 'src/analysis/analysis.service'
+import { GithubService } from './github.service';
 @Controller('github')
 export class GithubWebhookController {
-  constructor(private readonly githubWebhookService: GithubWebhookService) {}
+  constructor(
+    private readonly githubWebhookService: GithubWebhookService,
+    private readonly githubService: GithubService,
+    private readonly analysisService: AnalysisService,
+  ) {}
 
   @Post('webhook')
-  async webhook(@Body() body: any) {
-    console.log('webhook', body);
-
-    const results: any[] = [];
-
-    // Push 이벤트인지 확인
-    if (body.commits && Array.isArray(body.commits)) {
-      const repositoryName = body.repository?.full_name || 'unknown';
-
-      // 각 커밋을 순차적으로 처리
-      for (const commit of body.commits) {
-        const commitInfo = {
-          repository: repositoryName,
-          commitMessage: commit.message,
-          authorName: commit.author?.name || 'unknown',
-          commitUrl: commit.url,
-          changedFiles: {
-            added: commit.added || [],
-            modified: commit.modified || [],
-            removed: commit.removed || [],
-          },
-          commitHash: commit.id,
-          timestamp: commit.timestamp,
-        };
-
-        console.log('추출된 커밋 정보:', commitInfo);
-
-        try {
-          // 서비스로 커밋 정보 전달하고 결과 수집
-          const result =
-            await this.githubWebhookService.processCommit(commitInfo);
-          results.push({
-            commitHash: commit.id,
-            ...result,
-          });
-        } catch (error) {
-          console.error(`커밋 ${commit.id} 처리 중 오류:`, error);
-          results.push({
-            commitHash: commit.id,
-            success: false,
-            error: error.message,
-          });
-        }
+  async webhook(
+    @Body() body: any, 
+    @Headers() headers: any
+  ) {
+      // console.log('----------------------');
+      // console.log('headers', headers);
+      // console.log('webhook', body);
+      // console.log('----------------------');
+      const type = headers['x-github-event'];
+      if (type == 'push') {
+        return this.handlePushEvent(body);
+      } else if (type == 'pull-request') {
+        return this.handlePullRequestEvent(body);
       }
-    }
-
-    return {
-      status: 'success',
-      processedCommits: results.length,
-      results: results,
-    };
   }
 
   @Get('webhook/commit/:issueId')
@@ -90,7 +58,7 @@ export class GithubWebhookController {
     @Query('limit') limit: string = '10',
   ) {
     const limitNum = parseInt(limit, 10) || 10;
-
+    console.log("areyoucall")
     if (limitNum < 1 || limitNum > 100) {
       throw new Error('limit은 1 이상 100 이하여야 합니다.');
     }
@@ -100,5 +68,139 @@ export class GithubWebhookController {
       cursor,
       limitNum,
     );
+  }
+
+  private async handlePushEvent(body: any) {
+    const results: any[] = [];
+
+    // Push 이벤트인지 확인
+    if (body.commits && Array.isArray(body.commits)) {
+      const repositoryName = body.repository?.full_name || 'unknown';
+
+      // 각 커밋을 순차적으로 처리
+      for (const commit of body.commits) {
+        const commitInfo = {
+          repository: repositoryName,
+          commitMessage: commit.message,
+          authorName: commit.author?.name || 'unknown',
+          commitUrl: commit.url,
+          changedFiles: {
+            added: commit.added || [],
+            modified: commit.modified || [],
+            removed: commit.removed || [],
+          },
+          commitHash: commit.id,
+          timestamp: commit.timestamp,
+        };
+        try{
+        } catch{
+          console.error("커밋 실패, id: ", commit.commitHash);
+        }
+
+        console.log('추출된 커밋 정보:', commitInfo);
+
+        try {
+          // 서비스로 커밋 정보 전달하고 결과 수집
+          const result = await this.githubWebhookService.processCommit(commitInfo);
+          results.push({
+            commitHash: commit.id,
+            ...result,
+          });
+        } catch (error) {
+          console.error(`커밋 ${commit.id} 처리 중 오류:`, error);
+          results.push({
+            commitHash: commit.id,
+            success: false,
+            error: error.message,
+          });
+        }
+      }
+    }
+
+    return {
+      status: 'success',
+      processedCommits: results.length,
+      results: results,
+    };
+  }
+  // pull-request 
+  private async handlePullRequestEvent(body: any) {
+    const results: any[] = [];
+
+    if (body.pull_request) {
+      const repositoryName = body.repository?.full_name || 'unknown';
+      const [owner, repo] = repositoryName.split('/');
+      const { number, head, base, title, user } = body.pull_request;
+      const action = body.action;
+
+      console.log(`PR #${number}: ${title} (${action})`);
+
+      try {
+        // PR의 변경된 파일들을 가져오기
+        const changedFiles =
+          await this.githubService.getFilesChangedInPullRequest(
+            owner,
+            repo,
+            number,
+            body.sender?.id || 'unknown',
+          );
+
+        // C/C++ 파일만 필터링
+        const cppFiles = changedFiles.filter(
+          (file) => file.language === 'c' || file.language === 'cpp',
+        );
+
+        if (cppFiles.length > 0) {
+          // 코드 분석 수행
+          const analysisResults = await this.analysisService.analyzeFiles(
+            cppFiles.map((file) => ({
+              filename: file.filename,
+              content: file.content,
+              language: file.language as 'c' | 'cpp',
+            })),
+          );
+
+          const prInfo = {
+            repository: repositoryName,
+            pullRequestNumber: number,
+            title: title,
+            author: user?.login || 'unknown',
+            action: action,
+            headSha: head?.sha,
+            baseSha: base?.sha,
+            changedFiles: {
+              total: changedFiles.length,
+              cppFiles: cppFiles.length,
+            },
+            analysisResults: analysisResults,
+          };
+
+          console.log('분석 완료된 PR 정보:', prInfo);
+          results.push(prInfo);
+        } else {
+          console.log('C/C++ 파일이 없어서 분석을 건너뜁니다.');
+          results.push({
+            repository: repositoryName,
+            pullRequestNumber: number,
+            title: title,
+            action: action,
+            message: 'No C/C++ files to analyze',
+          });
+        }
+      } catch (error) {
+        console.error('PR 분석 중 오류 발생:', error);
+        results.push({
+          pullRequestNumber: number,
+          error: error.message,
+        });
+      }
+    }
+
+    return {
+      status: 'success',
+      eventType: 'pull_request',
+      processedCommits: results.length,
+      results: results,
+    };
   }
 }
